llama hf tgi (#2476)
* Update notebook

* update
