+++ sdk/python/foundation-models/system/inference/text-generation/llama-files/score/hf-tgi/scoring.json
@@ -1,0 +1,10 @@
+{
+    "client_init_args": {
+        "TIMEOUT": 90.0
+    },
+    "model_init_args": {
+        "MODEL_ID": "mlflow_model_folder/data/model",
+        "MLMODEL_PATH": "mlflow_model_folder/MLmodel",
+        "NUM_SHARD": "1"
+    }
+}
